Added arithmetic operations.
For now, the player absorbs them and they disappear. For different gameplay purposes I might introduce operations that are not destroyed and can be re-used.

Code begins to be messy at this point, but I still have plenty of time to refactor later.
Need to refactor: The player and numbers need to have a common "Sprite" object that shows the number based on the hold value.